Added cleanup modal to our dashboard, which will request a batch process to remove orphaned submissions.

+++ includes/Admin/Processes/DataCleanup.php
@@ -3,13 +3,12 @@
 /**
  * Class NF_Abstracts_Batch_Process
  */
-class NF_Admin_DataCleanup extends NF_Abstracts_BatchProcess
+class NF_Admin_Processes_DataCleanup extends NF_Abstracts_BatchProcess
 {
     private $response = array(
         'batch_complete' => false
     );
     protected $delete = array();
-    global $wpdb;
     
     /**
      * Constructor
@@ -29,6 +28,7 @@ class NF_Admin_DataCleanup extends NF_Abstracts_BatchProcess
      */
     public function process()
     {
+        global $wpdb;
         // If we've not already started the cleanup process...
         if ( ! get_option( 'nf_doing_data_cleanup' ) ) {
             // Run the startup process.
@@ -44,7 +44,7 @@ class NF_Admin_DataCleanup extends NF_Abstracts_BatchProcess
             // Fetch the last item on it.
             $id = array_pop( $this->delete );
             // Get a list of post IDs to delete.
-            $sql = "SELECT DISTINCT(`id`) FROM `" . $wpdb->prefix . "posts` WHERE `id` IN( SELECT DISTINCT(`post_id`) FROM `" . $wpdb->prefix . "postmeta` WHERE `meta_key` = '_form_id' AND `meta_value` = '0' ) AND `post_type` = 'nf_sub' LIMIT 500";
+            $sql = "SELECT DISTINCT(`id`) FROM `" . $wpdb->prefix . "posts` WHERE `id` IN( SELECT DISTINCT(`post_id`) FROM `" . $wpdb->prefix . "postmeta` WHERE `meta_key` = '_form_id' AND `meta_value` = '" . $id . "' ) AND `post_type` = 'nf_sub' LIMIT 500";
             $result = $wpdb->get_results( $sql, 'ARRAY_A' );
             // If we got 500 or more results...
             if ( 500 == count( $result ) ) {
@@ -54,15 +54,20 @@ class NF_Admin_DataCleanup extends NF_Abstracts_BatchProcess
             // Convert our results to something we can use in a query.
             array_walk( $result, array( $this, 'smush_results' ) );
             $sub_sql = implode( ', ', $result );
-            // Delete postmeta data.
-            $sql = "DELETE FROM `" . $wpdb->prefix . "postmeta` WHERE `post_id` IN(" . $sub_sql . ")";
-            $wpdb->query( $sql );
-            // Delete post data.
-            $sql = "DELETE FROM `" . $wpdb->prefix . "posts` WHERE `id` IN(" . $sub_sql . ")";
-            $wpdb->query( $sql );
+            // If we have something to query...
+            if ( '' != $sub_sql ) {
+                // Delete postmeta data.
+                $sql = "DELETE FROM `" . $wpdb->prefix . "postmeta` WHERE `post_id` IN(" . $sub_sql . ")";
+                $wpdb->query( $sql );
+                // Delete post data.
+                $sql = "DELETE FROM `" . $wpdb->prefix . "posts` WHERE `id` IN(" . $sub_sql . ")";
+                $wpdb->query( $sql );
+            }
         }
         // If our array isn't empty...
         if ( ! empty( $this->delete ) ) {
+            // Determine how many steps we have left.
+            $this->response[ 'step_remaining' ] = count( $this->delete );
             update_option( 'nf_data_cleanup_ids', implode( ',', $this->delete ) );
             echo wp_json_encode( $this->response );
             wp_die();
@@ -79,6 +84,7 @@ class NF_Admin_DataCleanup extends NF_Abstracts_BatchProcess
      */
     public function startup()
     {
+        global $wpdb;
         // Get a list of IDs from the forms table.
         $sql = "SELECT DISTINCT(`id`) FROM `" . $wpdb->prefix . "nf3_forms`";
         $forms = $wpdb->get_results( $sql, 'ARRAY_A' );
@@ -93,8 +99,10 @@ class NF_Admin_DataCleanup extends NF_Abstracts_BatchProcess
                 $this->delete[] = $form[ 'id' ];
             }
         }
+        // Get our number of steps for the progress bar.
+        $this->response[ 'step_total' ] = count( $this->delete );
         // Flag startup done.
-        add_option( 'nf_doing_data_clenaup', 'true' );
+        add_option( 'nf_doing_data_cleanup', 'true' );
     }
 
 
@@ -103,19 +111,21 @@ class NF_Admin_DataCleanup extends NF_Abstracts_BatchProcess
      */
     public function cleanup()
     {
+        global $wpdb;
         // Delete our options.
         delete_option( 'nf_data_cleanup_ids' );
         delete_option( 'nf_doing_data_cleanup' );
         // Add our "finished" option.
         add_option( 'ninja_forms_data_is_clean', 'true' );
         // Tell our JS that we're done.
+        $this->response[ 'step_remaining' ] = 0;
         $this->response[ 'batch_complete' ] = true;
     }
     
     /**
      * Function to compress data array and eliminate unnecessary keys.
      */
-    public function smush_results( $value, $key ) {
+    public function smush_results( &$value, $key ) {
         $value = $value[ 'id' ];
     }
 }
